BACKPORT: l2tp: fix race in l2tp_recv_common()
Taking a reference on sessions in l2tp_recv_common() is racy; this
has to be done by the callers.

To this end, a new function is required (l2tp_session_get()) to
atomically lookup a session and take a reference on it. Callers then
have to manually drop this reference.

Fixes: fd558d186df2 ("l2tp: Split pppol2tp patch into separate l2tp and ppp parts")
Signed-off-by: Guillaume Nault <g.nault@alphalink.fr>
Signed-off-by: David S. Miller <davem@davemloft.net>

(cherry picked from commit 61b9a047729bb230978178bca6729689d0c50ca2)
Signed-off-by: Daniel Rosenberg <drosen@google.com>
Bug: 38159931
Change-Id: I206a40ff51b4c8230d68acc2932f3277524d0f90
Signed-off-by: celtare21 <celtare21@gmail.com>